https://github.com/opencart/opencart/issues/6225

+++ upload/admin/model/upgrade/1000.php
@@ -77,7 +77,9 @@ class ModelUpgrade1000 extends Model {
 				$query = $this->db->query("SHOW INDEXES FROM `" . DB_PREFIX . $table['name'] . "`");
 
 				foreach ($query->rows as $result) {
-					$this->db->query("ALTER TABLE `" . DB_PREFIX . $table['name'] . "` DROP INDEX `" . $result['Key_name'] . "`");
+					if ($result['Key_name'] != 'PRIMARY') {
+						$this->db->query("ALTER TABLE `" . DB_PREFIX . $table['name'] . "` DROP INDEX `" . $result['Key_name'] . "`");
+					}
 				}
 
 				// Primary Key
